refactor: move out not strictly p2p tunnel stuff
