+++ src-tauri/src/network/mod.rs
@@ -11,8 +11,10 @@ use tokio::sync::oneshot;
 pub use user_identification::UserIdentification;
 
 pub mod ctrl_chnl;
-mod holepunch;
 mod p2p_tunl;
+mod resolver;
+mod rhizome;
+mod tls_kcp;
 mod user_identification;
 
 type ConnectionMap = HashMap<RoomId, Connection>;
